A bunch of little changes, plus the addition of gbpFITS.

+++ gbpLib/gbpParse/count_lines.c
@@ -10,10 +10,10 @@
 #include <stdio.h>
 #include <gbpCommon.h>
 int count_lines(FILE *fp){
-  int   n_lines=0;
-  char *line   =NULL;
-  int   n      =0;
-  int   r;
+  int     n_lines=0;
+  char   *line   =NULL;
+  size_t  n     =0;
+  int     r;
   while(!feof(fp)){
     r=getline(&line,&n,fp);
     if(r>0) n_lines++;
