+++ gbpLib/gbpParse/count_lines_data.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 #include <gbpCommon.h>
 int count_lines_data(FILE *fp){
-  int   n_lines=0;
-  char *line   =NULL;
-  int   n      =0;
-  int   r;
+  int     n_lines=0;
+  char   *line   =NULL;
+  size_t  n      =0;
+  int     r;
   while(!feof(fp)){
     r=getline(&line,&n,fp);
     if(r>0 && !check_comment(line))
